# docs/api/ai-api.openapi.json
{
  "components": {
    "schemas": {
...
      "BriefEvent": {
        "properties": {
          "brief": {
            "$ref": "#/components/schemas/TripBrief"
          },
          "missing": {
            "items": {
              "enum": [
                "destination",
                "origin",
                "travellers",
                "dates",
                "interests"
              ],
              "type": "string"
            },
            "title": "Missing",
            "type": "array"
          },
          "type": {
            "const": "brief",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "brief",
          "missing"
        ],
        "title": "BriefEvent",
        "type": "object"
      },
      "CardDetail": {
        "description": "One card in full (`GET /planner/card?id=`): what the panel shows when a\ntraveller opens an activity, on top of everything the carousel card has.\n\nAdditive over `OptionCard` and read from the same place \u2014 the corpus\ndocument the id names \u2014 so the client never sends content, only an id.\nLike every planner model, no field has a default: an unknown one travels\nas `null` and the generated TypeScript keeps it required.",
        "properties": {
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "category": {
            "title": "Category",
            "type": "string"
          },
          "deep_link": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deep Link"
          },
          "description": {
            "description": "The corpus document's own text, trimmed at a sentence boundary; empty when the document carries none",
            "title": "Description",
            "type": "string"
          },
          "district": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "District"
          },
          "heading_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Where the document sits in the city's corpus",
            "title": "Heading Path"
          },
          "hours": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hours"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "image_credit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Credit"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "lat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lat"
          },
          "license": {
            "title": "License",
            "type": "string"
          },
          "lon": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lon"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "price_tier": {
            "anyOf": [
              {
                "enum": [
                  1,
                  2,
                  3
                ],
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Tier"
          },
          "rating_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating Text"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "source_url": {
            "title": "Source Url",
            "type": "string"
          },
          "subtitle": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subtitle"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "website": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "The venue's own site, when it is not the source page",
            "title": "Website"
          },
          "why": {
            "maxLength": 140,
            "title": "Why",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "subtitle",
          "district",
          "category",
          "image_url",
          "image_credit",
          "price_tier",
          "rating_text",
          "hours",
          "lat",
          "lon",
          "why",
          "source",
          "source_url",
          "license",
          "deep_link",
          "description",
          "address",
          "phone",
          "website",
          "heading_path"
        ],
        "title": "CardDetail",
        "type": "object"
      },
      "ChatMessage": {
        "description": "A single message replayed from the conversation history.",
        "properties": {
          "content": {
            "description": "Message content",
            "maxLength": 8000,
            "title": "Content",
            "type": "string"
          },
          "role": {
            "description": "Message role: 'user' or 'assistant'",
            "enum": [
              "user",
              "assistant"
            ],
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "role",
          "content"
        ],
        "title": "ChatMessage",
        "type": "object"
      },
...
      "CityIntro": {
        "description": "A city's own description in one language: the lead of its Wikivoyage\narticle (CC BY-SA 4.0), trimmed at a sentence boundary, and the page it\ncomes from \u2014 shown with the text, as the licence asks.",
        "properties": {
          "source_url": {
            "title": "Source Url",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text",
          "source_url"
        ],
        "title": "CityIntro",
        "type": "object"
      },
...
      "DoneEvent": {
        "properties": {
          "type": {
            "const": "done",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type"
        ],
        "title": "DoneEvent",
        "type": "object"
      },
      "ErrorEvent": {
        "properties": {
          "error": {
            "title": "Error",
            "type": "string"
          },
          "error_code": {
            "title": "Error Code",
            "type": "string"
          },
          "type": {
            "const": "error",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "error",
          "error_code"
        ],
        "title": "ErrorEvent",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ItineraryPatchEvent": {
        "properties": {
          "ops": {
            "items": {
              "discriminator": {
                "mapping": {
                  "put_activity": "#/components/schemas/PutActivityOp",
                  "remove_activity": "#/components/schemas/RemoveActivityOp",
                  "set_day_title": "#/components/schemas/SetDayTitleOp",
                  "set_route": "#/components/schemas/SetRouteOp",
                  "set_stay": "#/components/schemas/SetStayOp",
                  "set_weather": "#/components/schemas/SetWeatherOp",
                  "warn": "#/components/schemas/WarnOp"
                },
                "propertyName": "op"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/SetStayOp"
                },
                {
                  "$ref": "#/components/schemas/PutActivityOp"
                },
                {
                  "$ref": "#/components/schemas/RemoveActivityOp"
                },
                {
                  "$ref": "#/components/schemas/SetDayTitleOp"
                },
                {
                  "$ref": "#/components/schemas/SetRouteOp"
                },
                {
                  "$ref": "#/components/schemas/SetWeatherOp"
                },
                {
                  "$ref": "#/components/schemas/WarnOp"
                }
              ]
            },
            "title": "Ops",
            "type": "array"
          },
          "type": {
            "const": "itinerary_patch",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "ops"
        ],
        "title": "ItineraryPatchEvent",
        "type": "object"
      },
      "ItinerarySnapshot": {
        "description": "What the client holds: the stay and the days, as card ids.",
        "properties": {
          "days": {
            "items": {
              "$ref": "#/components/schemas/DaySnapshot"
            },
            "maxItems": 14,
            "title": "Days",
            "type": "array"
          },
          "stay_card_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stay Card Id"
          }
        },
        "required": [
          "stay_card_id",
          "days"
        ],
        "title": "ItinerarySnapshot",
        "type": "object"
      },
...
      "PlannerCity": {
        "description": "A city the planner covers (`GET /planner/cities`): what the page needs\nto offer it as a destination and to introduce it on a trip overview. The\nspellings it answers to stay server-side.",
        "properties": {
          "centre": {
            "description": "[latitude, longitude]",
            "maxItems": 2,
            "minItems": 2,
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "title": "Centre",
            "type": "array"
          },
          "country": {
            "title": "Country",
            "type": "string"
          },
          "country_code": {
            "description": "ISO 3166-1 alpha-2, upper-case",
            "title": "Country Code",
            "type": "string"
          },
          "image_credit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "The line to print beside the photo (author and licence)",
            "title": "Image Credit"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "The city's photo on Wikimedia Commons, 1200 px wide",
            "title": "Image Url"
          },
          "intro": {
            "additionalProperties": {
              "$ref": "#/components/schemas/CityIntro"
            },
            "description": "The city's description keyed by language code (`en`, `es`); a language the corpus has no lead for is absent",
            "title": "Intro",
            "type": "object"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "name",
          "country",
          "country_code",
          "centre",
          "timezone",
          "intro",
          "image_url",
          "image_credit"
        ],
        "title": "PlannerCity",
        "type": "object"
      },
      "PlannerEvent": {
        "discriminator": {
          "mapping": {
            "brief": "#/components/schemas/BriefEvent",
            "done": "#/components/schemas/DoneEvent",
            "error": "#/components/schemas/ErrorEvent",
            "itinerary_patch": "#/components/schemas/ItineraryPatchEvent",
            "options": "#/components/schemas/OptionsEvent",
            "progress": "#/components/schemas/ProgressEvent",
            "text": "#/components/schemas/TextEvent"
          },
          "propertyName": "type"
        },
        "oneOf": [
          {
            "$ref": "#/components/schemas/TextEvent"
          },
          {
            "$ref": "#/components/schemas/BriefEvent"
          },
          {
            "$ref": "#/components/schemas/OptionsEvent"
          },
          {
            "$ref": "#/components/schemas/ItineraryPatchEvent"
          },
          {
            "$ref": "#/components/schemas/ErrorEvent"
          },
          {
            "$ref": "#/components/schemas/DoneEvent"
          },
          {
            "$ref": "#/components/schemas/ProgressEvent"
          }
        ]
      },
      "PlannerTurn": {
        "description": "One turn: a message, a structured action, or both.",
        "properties": {
          "action": {
            "anyOf": [
              {
                "discriminator": {
                  "mapping": {
                    "remove": "#/components/schemas/RemoveAction",
                    "select": "#/components/schemas/SelectAction"
                  },
                  "propertyName": "type"
                },
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/SelectAction"
                  },
                  {
                    "$ref": "#/components/schemas/RemoveAction"
                  }
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Action"
          },
          "brief": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TripBrief"
              },
              {
                "type": "null"
              }
            ]
          },
          "exclude_card_ids": {
            "description": "ids the traveller already saw for this ask; never offered again",
            "items": {
              "maxLength": 200,
              "minLength": 1,
              "type": "string"
            },
            "maxItems": 60,
            "title": "Exclude Card Ids",
            "type": "array"
          },
          "history": {
            "description": "Text turns before this one, oldest first",
            "items": {
              "$ref": "#/components/schemas/ChatMessage"
            },
            "maxItems": 40,
            "title": "History",
            "type": "array"
          },
          "itinerary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ItinerarySnapshot"
              },
              {
                "type": "null"
              }
            ]
          },
          "language": {
            "description": "The page's language: what the planner answers in unless the traveller's latest messages clearly read as the other one (TRA-246)",
            "enum": [
              "en",
              "es"
            ],
            "title": "Language",
            "type": "string"
          },
          "message": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "session_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "The planner draft this turn belongs to, minted by the page: links the turn's trace to the others of the same conversation (ADR 0024)",
            "title": "Session Id"
          },
          "trip_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "The Trip this draft was saved as, once it was (else null)",
            "title": "Trip Id"
          }
        },
        "required": [
          "message",
          "action",
          "history",
          "brief",
          "itinerary",
          "exclude_card_ids",
          "trip_id",
          "session_id",
          "language"
        ],
        "title": "PlannerTurn",
        "type": "object"
      },
...
      "RemoveAction": {
        "description": "An activity taken out of a slot.",
        "properties": {
          "card_id": {
            "minLength": 1,
            "title": "Card Id",
            "type": "string"
          },
          "slot": {
            "$ref": "#/components/schemas/Slot"
          },
          "type": {
            "const": "remove",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "slot",
          "card_id"
        ],
        "title": "RemoveAction",
        "type": "object"
      },
...
      "SelectAction": {
        "description": "Cards chosen in a carousel the server sent earlier (`group_id`).\n\nA group whose id names a slot (`slot:<day>:<part>`) places the pick on\nits own. A `found:` group does not: its cards arrived unplaced, so the\npage names the day and the part the traveller chose in `slot` (TRA-185,\nADR 0018). `null` for every placed group.",
        "properties": {
          "card_ids": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "minItems": 1,
            "title": "Card Ids",
            "type": "array"
          },
          "group_id": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Group Id",
            "type": "string"
          },
          "slot": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Slot"
              },
              {
                "type": "null"
              }
            ]
          },
          "type": {
            "const": "select",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "group_id",
          "card_ids",
          "slot"
        ],
        "title": "SelectAction",
        "type": "object"
      },
...
      "TextEvent": {
        "properties": {
          "delta": {
            "title": "Delta",
            "type": "string"
          },
          "type": {
            "const": "text",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "delta"
        ],
        "title": "TextEvent",
        "type": "object"
      },
...
      "TripBrief": {
        "description": "What the planner knows about the trip; the client sends it back each turn.",
        "properties": {
          "adults": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Adults"
          },
          "budget_tier": {
            "anyOf": [
              {
                "enum": [
                  1,
                  2,
                  3
                ],
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Tier"
          },
          "children": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Children"
          },
          "destination": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Destination"
          },
          "end_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "interests": {
            "items": {
              "type": "string"
            },
            "title": "Interests",
            "type": "array"
          },
          "nights": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nights"
          },
          "origin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin"
          },
          "pace": {
            "anyOf": [
              {
                "enum": [
                  "relaxed",
                  "balanced",
                  "intense"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pace"
          },
          "start_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          }
        },
        "required": [
          "destination",
          "origin",
          "start_date",
          "end_date",
          "nights",
          "adults",
          "children",
          "budget_tier",
          "interests",
          "pace"
        ],
        "title": "TripBrief",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Kyrian World \u2014 AI API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/ai/planner": {
      "post": {
        "description": "Stream the planner's answer to one turn for the authenticated user.\n\nThe service keeps no state: the turn carries the brief, the itinerary\nsnapshot and the transcript. Events are discriminated on `type` (`text`,\n`brief`, `options`, `itinerary_patch`, `progress`, `error`, `done`); the\nstream ends with `data: [DONE]`. Needs retrieval (`RETRIEVAL_ENABLED`):\nevery card is a corpus document, so without a store the endpoint answers\n503. The turn's trace is written to the interactions table before\n`[DONE]` (ADR 0024).\n\nWith `ACCESS_CONTROL_ENABLED` the turn is refused before the stream starts,\nas a plain JSON error: 403 `ACCESS_DENIED` for an account that is not on\nthe access list, 429 `DAILY_TOKEN_LIMIT` (`extras`: `limit`, `used`,\n`resets_at`) once today's tokens are spent (ADR 0026).",
        "operationId": "planner_api_v1_ai_planner_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlannerTurn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              },
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/PlannerEvent"
                }
              }
            },
            "description": "Server-Sent Events: one `PlannerEvent` JSON object per `data:` line, `data: [DONE]` last."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Planner",
        "tags": [
          "Planner"
        ]
      }
    },
    "/api/v1/ai/planner/card": {
      "get": {
        "description": "One card in full: the article behind it, its address, phone and site.\n\nThe id travels as a query parameter because corpus ids contain slashes\nand colons. It is read back from the store, so the answer never depends\non what the client kept; an id the index does not hold is a 404. Needs\nretrieval (`RETRIEVAL_ENABLED`), like the planner itself. Traced like a\nturn (ADR 0024), written before the answer is sent.",
        "operationId": "card_api_v1_ai_planner_card_get",
        "parameters": [
          {
            "description": "The corpus document id the card carries, e.g. `osm:relation/13067`",
            "in": "query",
            "name": "id",
            "required": true,
            "schema": {
              "description": "The corpus document id the card carries, e.g. `osm:relation/13067`",
              "maxLength": 200,
              "minLength": 1,
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Card",
        "tags": [
          "Planner"
        ]
      }
    },
    "/api/v1/ai/planner/cities": {
      "get": {
        "description": "The cities the planner can plan, from the corpus manifest shipped with\nthe service: the page offers them as destinations.",
        "operationId": "cities_api_v1_ai_planner_cities_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlannerCity"
                  },
                  "title": "Response Cities Api V1 Ai Planner Cities Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Cities",
        "tags": [
          "Planner"
        ]
      }
    },
...
  }
}